Moved insertNew user function to routes/authorization file
